Changed all to 64 bits (long int mode) now, the example with a : c = 11 + d - 7 works

+++ my-gcc/tests/test.h
@@ -1,1 +1,1 @@
-extern int falsemain(void);
+extern long int falsemain(void);
